Crate axiom_eth

Modules

block_header

keccak

mpt

providers

rlp

storage

util

Structs

EthChip

EthConfig

Config shared for block header and storage proof circuits

Enums

Network

Traits

Field

Trait used to reduce verbosity with the declaration of the [FieldExt] trait and its repr.