Function axiom_eth::block_header::get_boundary_block_data

pub fn get_boundary_block_data<'v, F: Field + PrimeField>(
    ctx: &mut Context<'_, F>,
    gate: &impl GateInstructions<F>,
    chain: &[EthBlockHeaderTrace<'v, F>],
    num_blocks_minus_one: &AssignedValue<'v, F>
) -> ([AssignedValue<'v, F>; 2], [AssignedValue<'v, F>; 2], AssignedValue<'v, F>)

Given a block header chain chain of fixed length 2^max_depth, returns

(prev_block_hash, end_block_hash, start_block_number || end_block_number)

where

prev_block_hash = chain[0].parent_hash,
end_block_hash = chain[num_blocks_minus_one].block_hash,
start_block_number = chain[0].number,
end_block_number = chain[num_blocks_minus_one].number

The hashes are H256 that are represented as two u128 (we need them in 128 bits to fit in Bn254 scalar field).

The numbers are left padded by zeros to be exactly 4 bytes (u32); the two padded numbers are concatenated together to a u64.