Trait axiom_eth::block_header::EthBlockHeaderChip

pub trait EthBlockHeaderChip<'v, F: Field> {
    fn decompose_block_header_phase0(
        &mut self,
        ctx: &mut Context<'v, F>,
        block_header: &[u8],
        network: Network
    ) -> EthBlockHeaderTraceWitness<'v, F>;
    fn decompose_block_header_phase1(
        &mut self,
        ctx: &mut Context<'v, F>,
        witness: EthBlockHeaderTraceWitness<'v, F>
    ) -> EthBlockHeaderTrace<'v, F>;
    fn decompose_block_header_chain_phase1(
        &mut self,
        ctx: &mut Context<'v, F>,
        witnesses: Vec<EthBlockHeaderTraceWitness<'v, F>>,
        num_blocks_minus_one: Option<&AssignedValue<'v, F>>
    ) -> Vec<EthBlockHeaderTrace<'v, F>>;

    fn decompose_block_header_chain_phase0(
        &mut self,
        ctx: &mut Context<'v, F>,
        headers: &[Vec<u8>],
        network: Network
    ) -> Vec<EthBlockHeaderTraceWitness<'v, F>> { ... }
}

Required Methods

fn decompose_block_header_phase0(
&mut self,
ctx: &mut Context<'v, F>,
block_header: &[u8],
network: Network
) -> EthBlockHeaderTraceWitness<'v, F>

Takes the variable length RLP encoded block header, padded with 0s to the maximum possible block header RLP length, and outputs the decomposition into block header fields.

In addition, the keccak block hash of the block is calculated.

Assumes block_header and block_header_assigned have the same values as bytes. The former is only used for faster witness generation.

This is the preparation step that computes the witnesses. This MUST be done in FirstPhase. The accompanying decompose_block_header_finalize must be called in SecondPhase to constrain the RLCs associated to the RLP decoding.

fn decompose_block_header_phase1(
&mut self,
ctx: &mut Context<'v, F>,
witness: EthBlockHeaderTraceWitness<'v, F>
) -> EthBlockHeaderTrace<'v, F>

Takes the variable length RLP encoded block header, padded with 0s to the maximum possible block header RLP length, and outputs the decomposition into block header fields.

In addition, the keccak block hash of the block is calculated.

Assumes block_header and block_header_assigned have the same values as bytes. The former is only used for faster witness generation.

This is the finalization step that constrains RLC concatenations. This should be called after decompose_block_header_phase0. This MUST be done in SecondPhase.

fn decompose_block_header_chain_phase1(
&mut self,
ctx: &mut Context<'v, F>,
witnesses: Vec<EthBlockHeaderTraceWitness<'v, F>>,
num_blocks_minus_one: Option<&AssignedValue<'v, F>>
) -> Vec<EthBlockHeaderTrace<'v, F>>

Takes a list of 2^max_depth (purported) RLP encoded block headers. Decomposes each header into it’s fields. headers[0] is the earliest block

If num_blocks_minus_one is not None, then the circuit checks that the first num_blocks := num_blocks_minus_one + 1 block headers form a chain: meaning that the parent hash of block i + 1 equals the hash of block i.

Otherwise if num_blocks is None, the circuit checks that all headers form a hash chain.

Assumes that 0 <= num_blocks_minus_one < 2^max_depth.

This is the finalization step that constrains RLC concatenations. In this step the hash chain is actually constrained. This should be called after decompose_block_header_chain_phase0. This MUST be done in SecondPhase.

Provided Methods

fn decompose_block_header_chain_phase0(
&mut self,
ctx: &mut Context<'v, F>,
headers: &[Vec<u8>],
network: Network
) -> Vec<EthBlockHeaderTraceWitness<'v, F>>

Takes a list of (purported) RLP encoded block headers and decomposes each header into it’s fields. headers[0] is the earliest block.

This is the preparation step that computes the witnesses. This MUST be done in FirstPhase. The accompanying decompose_block_header_chain_phase1 must be called in SecondPhase to constrain the RLCs associated to the RLP decoding.