Module axiom_eth::rlp::rlc

Structs

RlcChip

This chip is a wrapper around RlcConfig together with cached assigned values of the powers of the challenge gamma. The chip can be mutably borrowed so that the cache can be updated to higher powers.

RlcConfig

This config consists of a variable number of advice columns, all in SecondPhase. Each advice column has a selector column that enables a custom gate to aid RLC computation.

RlcFixedTrace

RLC of a trace of known fixed length

RlcTrace

RLC of a trace of variable length but known maximum length

RlcVarLen

RlcVarRef

Constants

RLC_PHASE

Functions

rlc_constrain_equal

rlc_is_equal

Define the dynamic RLC: RLC(a, l) = \sum_{i = 0}^{l - 1} a_i r^{l - 1 - i} where a is a variable length vector of length l.

rlc_select

rlc_select_from_idx