Module axiom_eth::util

Structs

EthConfigParams

Functions

bytes_be_to_u128

Assumes that bytes have witnesses that are bytes.

bytes_be_to_uint

bytes_be_var_to_fixed

Takes a fixed length array bytes and returns a length out_len array equal to [[0; out_len - len], bytes[..len]].concat(), i.e., we take bytes[..len] and zero pad it on the left.

decode_field_to_addr

decode_field_to_h256

decode_field_to_u256

encode_addr_to_field

encode_h256_to_field

Takes hash as bytes32 and returns (hash[..16], hash[16..]) represented as big endian numbers in the prime field

encode_u256_to_field

Takes U256, converts to bytes32 (big endian) and returns (hash[..16], hash[16..]) represented as big endian numbers in the prime field

get_merkle_mountain_range

hash_tree_root

num_to_bytes_be

u256_to_bytes32_be

uint_to_bytes_be

uint_to_bytes_le